improve tests structure folders and improve tests
