+++ tests/emonpy/test_emonpy.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 from emon_tools.api_utils import SUCCESS_KEY
 from emon_tools.emonpy import EmonPy
-from tests.emon_api.emonpy_test_data import EmonpyDataTest as dtest
+from tests.emonpy.emonpy_test_data import EmonpyDataTest as dtest
 
 
 class TestEmonPy:
